Fix error-creating (invalid dates) data.sql and add datasource properties to fix idle system breakdown (see details below)
Simplistic method for creating dynamic dates for showings could produce invalid dates (like 2016-05-32). Now we just create test showings for 'today' (day of build/deploy). A bug was detected: the program would run fine for as long as we actively tested it, but if left deployed on server over night (with no action) the database connection would fail and result in a broken site the next day. The new datasource properties in application.yml fixes this. See see http://stackoverflow.com/questions/30451470/connection-to-db-dies-after-424-in-spring-boot-jpa-hibernate for futher explaination.